# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Benefits Enrollment MVP API
  version: 0.2.0
  description: Multi-tenant benefits enrollment API with role-based access controls.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
    Role:
      type: string
      enum: [FULL_ADMIN, COMPANY_ADMIN, EMPLOYEE]
    InviteTargetRole:
      type: string
      enum: [COMPANY_ADMIN, EMPLOYEE]
    BenefitClass:
      type: string
      enum: [FULL_TIME_ELIGIBLE, INELIGIBLE]
    EmploymentStatus:
      type: string
      enum: [ACTIVE, TERMED]
    PlanType:
      type: string
      enum: [MEDICAL, DENTAL, VISION]
    CoverageTier:
      type: string
      enum: [EMPLOYEE_ONLY, EMPLOYEE_SPOUSE, EMPLOYEE_CHILDREN, FAMILY]
...
    Tenant:
      type: object
      required: [id, companyId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InviteCode:
      type: object
      required: [id, tenantId, code, targetRole, createdByUserId, usesCount, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        code:
          type: string
        targetRole:
          $ref: '#/components/schemas/InviteTargetRole'
        createdByUserId:
          type: string
          format: uuid
        expiresAt:
          type: [string, 'null']
          format: date-time
        maxUses:
          type: [integer, 'null']
        usesCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    User:
      type: object
      required: [id, tenantId, email, role, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: [string, 'null']
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SecurityEvent:
      type: object
      required: [id, eventType, severity, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: [string, 'null']
          format: uuid
        tenantId:
          type: [string, 'null']
          format: uuid
        eventType:
          type: string
        severity:
          type: string
          enum: [INFO, WARN, ERROR]
        ipAddress:
          type: [string, 'null']
        userAgent:
          type: [string, 'null']
        metadata:
          type: [object, 'null']
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    EmployeeProfileInput:
      type: object
      required:
        - employeeId
        - firstName
        - lastName
        - dob
        - hireDate
        - salaryAmount
        - benefitClass
        - employmentStatus
      properties:
        employeeUserId:
          type: string
          format: uuid
          description: Required only for FULL_ADMIN calls to employee-scoped routes.
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dob:
          type: string
          format: date
        hireDate:
          type: string
          format: date
        salaryAmount:
          type: number
        benefitClass:
          $ref: '#/components/schemas/BenefitClass'
        employmentStatus:
          $ref: '#/components/schemas/EmploymentStatus'
    EmployeeProfile:
      allOf:
        - $ref: '#/components/schemas/EmployeeProfileInput'
        - type: object
          required: [userId, tenantId, createdAt, updatedAt]
          properties:
            userId:
              type: string
              format: uuid
            tenantId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    PlanYear:
      type: object
      required: [id, tenantId, name, startDate, endDate, createdByUserId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Plan:
      type: object
      required: [id, tenantId, planYearId, type, carrier, planName, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        planYearId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/PlanType'
        carrier:
          type: string
        planName:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlanPremiumInput:
      type: object
      required: [coverageTier, employeeMonthlyCost, employerMonthlyCost]
      properties:
        coverageTier:
          $ref: '#/components/schemas/CoverageTier'
        employeeMonthlyCost:
          type: number
          minimum: 0
        employerMonthlyCost:
          type: number
          minimum: 0
    PlanPremium:
      allOf:
        - $ref: '#/components/schemas/PlanPremiumInput'
        - type: object
          required: [id, planId, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            planId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        companyId:
          type: string
    CreateInviteCodeRequest:
      type: object
      properties:
        expiresAt:
          type: string
          format: date-time
        maxUses:
          type: integer
          minimum: 1
    CreatePlanYearRequest:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    CreatePlanRequest:
      type: object
      required: [planYearId, type, carrier, planName]
      properties:
        planYearId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/PlanType'
        carrier:
          type: string
        planName:
          type: string
    SetPlanPremiumsRequest:
      type: object
      required: [tiers]
      properties:
        tiers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PlanPremiumInput'
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: Invalid or expired token
...
paths:
...
  /full-admin/tenants:
    get:
      tags: [Full Admin]
      summary: List all tenants
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tenant list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Full Admin]
      summary: Create tenant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    $ref: '#/components/schemas/Tenant'
  /full-admin/tenants/{tenantId}/invite-codes/company-admin:
    post:
      tags: [Full Admin]
      summary: Create company admin invite code
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteCodeRequest'
      responses:
        '201':
          description: Invite code created
          content:
            application/json:
              schema:
                type: object
                properties:
                  inviteCode:
                    $ref: '#/components/schemas/InviteCode'
  /full-admin/security-events:
    get:
      tags: [Full Admin]
      summary: List security events across tenants
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Security events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecurityEvent'
  /tenants/{tenantId}/company-admin/users:
    get:
      tags: [Company Admin]
      summary: List tenant users
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [COMPANY_ADMIN, EMPLOYEE]
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /tenants/{tenantId}/company-admin/security-events:
    get:
      tags: [Company Admin]
      summary: List security events for tenant
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Tenant security events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecurityEvent'
  /tenants/{tenantId}/company-admin/invite-codes/employee:
    post:
      tags: [Company Admin]
      summary: Create employee invite code
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteCodeRequest'
      responses:
        '201':
          description: Invite code created
          content:
            application/json:
              schema:
                type: object
                properties:
                  inviteCode:
                    $ref: '#/components/schemas/InviteCode'
  /tenants/{tenantId}/company-admin/employees/{employeeUserId}/profile:
    put:
      tags: [Company Admin]
      summary: Upsert employee profile
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: employeeUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeProfileInput'
      responses:
        '200':
          description: Employee profile saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: '#/components/schemas/EmployeeProfile'
  /tenants/{tenantId}/company-admin/plan-years:
    get:
      tags: [Company Admin]
      summary: List plan years for tenant
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan year list
          content:
            application/json:
              schema:
                type: object
                properties:
                  planYears:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlanYear'
    post:
      tags: [Company Admin]
      summary: Create plan year
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanYearRequest'
      responses:
        '201':
          description: Plan year created
          content:
            application/json:
              schema:
                type: object
                properties:
                  planYear:
                    $ref: '#/components/schemas/PlanYear'
  /tenants/{tenantId}/company-admin/plans:
    get:
      tags: [Company Admin]
      summary: List plans for tenant
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: planYearId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan list
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plan'
    post:
      tags: [Company Admin]
      summary: Create plan
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Plan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/Plan'
  /tenants/{tenantId}/company-admin/plans/{planId}/premiums:
    put:
      tags: [Company Admin]
      summary: Replace plan premium tiers
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPlanPremiumsRequest'
      responses:
        '200':
          description: Premiums updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  premiums:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlanPremium'
...